# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Identizen index API
  version: 0.1.0
  summary: The Identizen index — device registry, login relay, and OpenID Provider.
  description: |
    Machine-readable description of the HTTP API served by an Identizen index
    (`apps/index`). It is written by hand from the Hono routes in
    `apps/index/src/routes/*.ts`, the Zod schemas in `packages/protocol/src/schemas.ts`,
    and the error helpers in `apps/index/src/lib/errors.ts`, and it is validated by
    `apps/index/test/openapi.test.ts`: every route registered on the app must appear here and
    every path here must be served by the app. The index also serves this document as JSON at
    `GET /openapi.json`.

    Three kinds of caller use this API:

    - **Phones** (the Identizen app): register with `POST /devices`, then sign every request
      with the `Idz-Signature` header (PROTOCOL.md §8).
    - **Browsers / the SDK**: start logins (`POST /challenge`), run discovery
      (`/discover/*`), and wait on the challenge (`/challenge/{id}/state` or the WebSocket).
    - **Sites** (relying parties): standard OIDC (`/authorize`, `/token`, `/userinfo`), site
      management (`/sites`), and the server-to-server Verification API (`/v1/verify`).

    Errors are JSON `{ "error": "<code>", "error_description": "…" }`; the codes are listed
    per operation and documented at https://docs.identizen.com/errors/. Unknown routes answer
...
    headers are `Content-Type`, `Authorization`, and `Idz-Signature`).
...
servers:
...
paths:
...
  /authorize:
    get:
      tags: [OIDC]
      operationId: authorize
      summary: OIDC authorization endpoint (hosted login page)
      description: |
        Authorization Code flow with PKCE. Creates a challenge session and renders the hosted
        login page (match code, QR, WebSocket to the session). After the phone approves, the
        page redirects the browser to `redirect_uri?code=…&state=…`. The code is single-use and
        bound to the client.

        Validation order: unknown `client_id` or unregistered `redirect_uri` is a JSON `400`
        (nothing is redirected to an unverified URI). Every later failure is an OIDC error
        redirect (`302` to `redirect_uri` with `error`, `error_description`, and `state`):
        `invalid_request` (missing `response_type`, missing or malformed PKCE,
        `acr_values=idz:mfa` alone without `login_hint`, `prompt=none` combined with another value),
        `unsupported_response_type`, `request_not_supported`, `request_uri_not_supported`,
        `invalid_scope` (no `openid`), `interaction_required` (`prompt=none`), and
        `login_required` (no active device is bound to `login_hint`). Unknown scopes and
        unknown `acr_values` are ignored (RFC 6749 §3.3; `acr` is a voluntary claim).

        `acr_values=idz:mfa` with `login_hint=<sub>` is step-up: the challenge is pushed to the
        device bound to `sub`. `login_hint` alone pushes to the bound device at `idz:login`.
        `prompt=enroll` marks an enrollment (the resulting `sub` is what the site stores).
        Rate limited per source IP (`429 rate_limited`) and per client (`429 client_rate_limited`); a step-up whose push would exceed the device's quota is `429 push_rate_limited`.
        The same request may be sent as `POST` with a form body (OpenID Connect Core §3.1.2.1).
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            const: code
        - name: client_id
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ClientId'
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: Must exactly match one of the site's registered `redirect_uris`.
        - name: scope
          in: query
          required: true
          schema:
            type: string
          description: Space-separated; must include `openid`. `handle` releases `idz_handle`.
          example: openid handle
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            const: S256
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: Echoed on the success and error redirects.
        - name: nonce
          in: query
          required: false
          schema:
            type: string
          description: Echoed in the id_token when sent.
        - name: acr_values
          in: query
          required: false
          schema:
            type: string
          description: Space-separated; each value must be `idz:login` or `idz:mfa`. `idz:mfa` requires `login_hint` unless `idz:login` is also listed, in which case the login proceeds at `idz:login`.
        - name: login_hint
          in: query
          required: false
          schema:
            type: string
          description: The per-site `sub` whose bound device should receive the push.
        - name: prompt
          in: query
          required: false
          schema:
            type: string
          description: '`none` always fails with `interaction_required`; `enroll` marks an enrollment. Other values are carried through unchanged.'
      responses:
        '200':
          $ref: '#/components/responses/LoginPage'
        '302':
          $ref: '#/components/responses/AuthorizeErrorRedirect'
        '400':
          $ref: '#/components/responses/AuthorizeBadRequest'
        '429':
          description: '`rate_limited` — per-IP limit; `client_rate_limited` — the site started too many challenges this minute; `push_rate_limited` — a step-up would push to a device that was pushed ten times in the last minute.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [OIDC]
      operationId: authorizePost
      summary: OIDC authorization endpoint (POST form)
      description: |
        Identical to `GET /authorize` with the same parameters in an
        `application/x-www-form-urlencoded` body (OpenID Connect Core §3.1.2.1 requires both).
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AuthorizeRequest'
      responses:
        '200':
          $ref: '#/components/responses/LoginPage'
        '302':
          $ref: '#/components/responses/AuthorizeErrorRedirect'
        '400':
          $ref: '#/components/responses/AuthorizeBadRequest'
        '429':
          description: '`rate_limited` — per-IP limit; `client_rate_limited` — the site started too many challenges this minute; `push_rate_limited` — a step-up would push to a device that was pushed ten times in the last minute.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /token:
    post:
      tags: [OIDC]
      operationId: token
      summary: OIDC token endpoint
      description: |
        `grant_type=authorization_code` with PKCE. Client authentication is
        `client_secret_basic` (HTTP Basic), `client_secret_post` (`client_id` + `client_secret`
        in the body), or `none` (`client_id` alone) for public clients; a confidential client
        that omits or mismatches its secret gets `401 invalid_client`. The code is redeemed once
        at the challenge session; `redirect_uri` must equal the one used on the authorization
        request; `code_verifier` must hash (S256) to the request's `code_challenge` (and must
        be absent when the request had none, which only `OIDC_PKCE_OPTIONAL=true` allows); the
        approving device must still be active. Each successful exchange creates a 30-day
        Identizen session (`sid`).
      security:
        - {}
        - clientSecretBasic: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Tokens. No refresh token is issued.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
            Pragma:
              schema:
                type: string
                const: no-cache
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: '`invalid_request` — `grant_type`, `code`, or `code_verifier` missing; `unsupported_grant_type` — any grant other than `authorization_code` (there are no refresh tokens); `invalid_grant` — malformed, unknown, expired (five minutes after approval), or used code, code issued to another client, `redirect_uri` mismatch or no longer registered, PKCE failure, or the device no longer exists / is not active. Every check runs before the code is consumed, so a failed exchange leaves the code usable by the right caller.'
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: '`invalid_client` — `client_id` missing, unknown, malformed Basic credentials, or the client secret is wrong. When the client used HTTP Basic the response carries `WWW-Authenticate: Basic realm="identizen"` (RFC 6749 §5.2).'
          headers:
            WWW-Authenticate:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /userinfo:
    get:
      tags: [OIDC]
      operationId: getUserinfo
      summary: OIDC userinfo endpoint
      description: |
        Requires the access token from `/token`. Fails with `401 invalid_token` once the
        session has been revoked, so it doubles as a session liveness check.
      security:
        - accessToken: []
      responses:
        '200':
          $ref: '#/components/responses/Userinfo'
        '401':
          $ref: '#/components/responses/InvalidToken'
    post:
      tags: [OIDC]
      operationId: postUserinfo
      summary: OIDC userinfo endpoint (POST form)
      description: Identical to `GET /userinfo`; the token travels in the `Authorization` header or, per RFC 6750 §2.2, as `access_token` in an `application/x-www-form-urlencoded` body.
      security:
        - accessToken: []
      responses:
        '200':
          $ref: '#/components/responses/Userinfo'
        '401':
          $ref: '#/components/responses/InvalidToken'
...
  /challenge:
    post:
      tags: [Challenges]
      operationId: startChallenge
      summary: Start a login (the JSON twin of `/authorize`)
      description: |
        What the SDK calls. Creates a signed challenge in a session and returns what the
        browser needs; the phone fetches the signed challenge itself (`GET /challenge/{id}`).
        When `redirect_uri` is present the session also carries the OIDC parameters so the
        approval yields an authorization code redeemable at `POST /token`; the request is then
        held to exactly the rules of `/authorize` (a registered `redirect_uri`, PKCE S256,
        `scope` including `openid`, `prompt` semantics), with errors returned as JSON
        (`400 invalid_request`, `invalid_scope`, `interaction_required`) instead of a redirect. `acr: idz:mfa` or
        `login_hint` pushes the challenge to the device bound to `login_hint`
        (`400 login_hint_required` when `idz:mfa` has no hint, `400 login_required` when the
        `sub` has no active bound device). Rate limited per source IP and per client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeStartRequest'
      responses:
        '201':
          description: Challenge created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeStarted'
        '400':
          description: '`invalid_request` — body failed validation, or (with `redirect_uri`) an authorization-request rule failed; `invalid_scope` — `scope` without `openid`; `interaction_required` — `prompt=none`; `login_hint_required`; `login_required` — no active device is bound to `login_hint` for this site.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: '`unknown_client` — no site with that `client_id`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: '`rate_limited` — per-IP limit; `client_rate_limited` — the site started too many challenges this minute; `push_rate_limited` — a step-up would push to a device that was pushed ten times in the last minute.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /challenge/{id}/ws:
    get:
      tags: [Challenges]
      operationId: challengeWebSocket
      summary: WebSocket for the waiting browser
      description: |
        Upgrade to a WebSocket bridged to the session. The server sends JSON text frames:
        `{ "type": "pending", "challenge_id", "exp" }` on connect while pending, then exactly one
        terminal event and a close (code 1000):
        `{ "type": "approved", "challenge_id", "pairing": SignedPairing | null, "redirect": string | null }`,
        `{ "type": "denied", "challenge_id" }`, or `{ "type": "expired", "challenge_id" }`.
        A late joiner receives the terminal event immediately. Sending the text `ping` returns
        `pong`. Without an `Upgrade: websocket` header the response is `426` (plain text).
      parameters:
        - $ref: '#/components/parameters/ChallengeIdPath'
        - name: Upgrade
          in: header
          required: true
          schema:
            type: string
            const: websocket
      responses:
        '101':
          description: Switching Protocols.
        '404':
          description: Unknown challenge (plain text `unknown challenge`).
          content:
            text/plain:
              schema:
                type: string
        '426':
          description: Not a WebSocket upgrade (plain text `expected websocket`).
          content:
            text/plain:
              schema:
                type: string
...
  /sites/{client_id}:
    get:
      tags: [Sites]
      operationId: getSite
      summary: Read a site
      description: Bearer token = the site's client secret (HTTP Basic with the secret as password is accepted too). Public clients have no secret and cannot authenticate here.
      security:
        - clientSecretBearer: []
        - clientSecretBasic: []
      parameters:
        - $ref: '#/components/parameters/ClientIdPath'
      responses:
        '200':
          description: The site.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
        '401':
          $ref: '#/components/responses/InvalidClient'
        '404':
          $ref: '#/components/responses/UnknownClient'
    patch:
      tags: [Sites]
      operationId: updateSite
      summary: Update a site, optionally rotating its secrets
      security:
        - clientSecretBearer: []
        - clientSecretBasic: []
      parameters:
        - $ref: '#/components/parameters/ClientIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SitePatch'
      responses:
        '200':
          description: The updated site. `client_secret` / `webhook_secret` are the new values when rotated, otherwise `null`.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Site'
                  - $ref: '#/components/schemas/SiteSecrets'
        '400':
          description: '`invalid_request` — body failed validation; `invalid_destination` — `webhook_url` or `backchannel_logout_uri` is one the index will not call (not https, credentials in the URL, or a local, private or link-local address).'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/InvalidClient'
        '404':
          $ref: '#/components/responses/UnknownClient'
      callbacks:
        backchannelLogout:
          $ref: '#/components/callbacks/BackchannelLogout'
        verificationWebhook:
          $ref: '#/components/callbacks/VerificationWebhook'
...
components:
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The OIDC access token from `POST /token` (`typ: at+jwt`, ES256, 1 hour, audience =
        the index). `/userinfo` accepts any client's token. `/me/*` additionally requires the
        token to have been issued to a client listed in the index's `DASHBOARD_CLIENT_IDS`
        (`403 not_dashboard_client`) and checks the session for revocation on every call.
    idzSignature:
      type: apiKey
      in: header
      name: Idz-Signature
      description: |
        Device request authentication (PROTOCOL.md §8):

            Idz-Signature: v1,d=<device_id>,t=<unix seconds>,s=<base64url sig>

        where `sig = Ed25519.sign(deviceKey, UTF8("identizen/v1/request\n" + METHOD + "\n" +
        PATH + "\n" + base64url(SHA-256(body)) + "\n" + t))`. `METHOD` is upper-case, `PATH`
        is the request path including the query string, `body` is the raw request body (empty
        string for none). The index rejects `t` outside ±60 s of its clock (`401 bad_signature`)
        and any `(device_id, t, sig)` seen before (`401 replayed_request`); a device may send at
        most 120 signed requests per minute. `@identizen/protocol` exports `signRequest`.
    clientSecretBearer:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <client_secret>`. On `/sites/{client_id}` the client id comes
        from the path; on `/v1/verify*` it must be sent in the `Idz-Client-Id` header.
    clientSecretBasic:
      type: http
      scheme: basic
      description: '`Authorization: Basic base64(client_id:client_secret)`.'
    siteRegistrationToken:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <SITE_REGISTRATION_TOKEN>`; required by `POST /sites` unless the index runs with `OPEN_SITE_REGISTRATION=true`.'
...
  responses:
...
    LoginPage:
      description: The hosted login page.
      headers:
        Cache-Control:
          schema:
            type: string
            const: no-store
        X-Frame-Options:
          schema:
            type: string
            const: DENY
        Referrer-Policy:
          schema:
            type: string
            const: no-referrer
      content:
        text/html:
          schema:
            type: string
    AuthorizeErrorRedirect:
      description: OIDC error redirect to `redirect_uri` with `error`, `error_description`, and `state` query parameters.
      headers:
        Location:
          schema:
            type: string
            format: uri
    AuthorizeBadRequest:
      description: '`invalid_client` — unknown `client_id`; `invalid_request` — `redirect_uri` is not registered for this client.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    ClientId:
      type: string
      description: '`idz_live_<ULID>` or `idz_test_<ULID>`.'
      pattern: '^idz_(live|test)_[0-9A-HJKMNP-TV-Z]{26}$'
...